// src/generated/graphql.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  AnalyticsDateTime: { input: any; output: any }
  AuctionsDateTime: { input: any; output: any }
  BigInt: { input: any; output: any }
  CommerceDate: { input: any; output: any }
  CommerceDateTime: { input: any; output: any }
  Date: { input: any; output: any }
  FormattedNumber: { input: any; output: any }
  ISO8601DateTime: { input: any; output: any }
  JSON: { input: any; output: any }
  Long: { input: any; output: any }
}
// ...
export type AnalyticsPartnerInquiryCountTimeSeriesStats = {
  __typename?: "AnalyticsPartnerInquiryCountTimeSeriesStats"
  count?: Maybe<Scalars["Int"]["output"]>
  endTime?: Maybe<Scalars["AnalyticsDateTime"]["output"]>
  startTime?: Maybe<Scalars["AnalyticsDateTime"]["output"]>
}

export type AnalyticsPartnerInquiryStats = {
  __typename?: "AnalyticsPartnerInquiryStats"
  inquiryCount: Scalars["Int"]["output"]
  inquiryResponseTime?: Maybe<Scalars["Int"]["output"]>
  partnerId: Scalars["String"]["output"]
  period: AnalyticsQueryPeriodEnum
  timeSeries?: Maybe<Array<AnalyticsPartnerInquiryCountTimeSeriesStats>>
}
// ...
export type AnalyticsPartnerStats = {
  __typename?: "AnalyticsPartnerStats"
  artworkPublished?: Maybe<AnalyticsPartnerStatsArtworksPublished>
  /** @deprecated Use artworkPublished for refactored time series bucket code */
  artworksPublished?: Maybe<AnalyticsArtworksPublishedStats>
  audience?: Maybe<AnalyticsPartnerAudienceStats>
  groupedStats?: Maybe<AnalyticsGroupedStatsConnection>
  inquiry?: Maybe<AnalyticsPartnerInquiryStats>
  pageview?: Maybe<AnalyticsPartnerStatsPageviews>
  /** @deprecated Use pageview for refactored time series bucket code */
  pageviews?: Maybe<AnalyticsPageviewStats>
  partnerId: Scalars["String"]["output"]
  rankedStats?: Maybe<AnalyticsRankedStatsConnection>
  sales?: Maybe<AnalyticsPartnerSalesStats>
  /** @deprecated Use rankedStats(objectType: ) instead */
  topArtworks?: Maybe<AnalyticsRankedStatsConnection>
  /** @deprecated Use audience() { uniqueVisitors } instead */
  uniqueVisitors?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type ArticleConnection = {
  __typename?: "ArticleConnection"
  edges?: Maybe<Array<Maybe<ArticleEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type ArtistConnection = {
  __typename?: "ArtistConnection"
  edges?: Maybe<Array<Maybe<ArtistEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type ArtistPartnerConnection = {
  __typename?: "ArtistPartnerConnection"
  edges?: Maybe<Array<Maybe<ArtistPartnerEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type ArtistsWithAlertCountsConnection = {
  __typename?: "ArtistsWithAlertCountsConnection"
  edges?: Maybe<Array<Maybe<ArtistsWithAlertCountsEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type ArtworkConnection = ArtworkConnectionInterface & {
  __typename?: "ArtworkConnection"
  edges?: Maybe<Array<Maybe<ArtworkEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type ArtworkImportConnection = {
  __typename?: "ArtworkImportConnection"
  edges?: Maybe<Array<Maybe<ArtworkImportEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type Contact = {
  __typename?: "Contact"
  canContact?: Maybe<Scalars["Boolean"]["output"]>
  email?: Maybe<Scalars["String"]["output"]>
  id: Scalars["ID"]["output"]
  internalID: Scalars["ID"]["output"]
  location?: Maybe<Location>
  name?: Maybe<Scalars["String"]["output"]>
  phone?: Maybe<Scalars["String"]["output"]>
  position?: Maybe<Scalars["String"]["output"]>
  slug: Scalars["ID"]["output"]
}

export type ContactConnection = {
  __typename?: "ContactConnection"
  edges?: Maybe<Array<Maybe<ContactEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type FilterArtworksConnection = ArtworkConnectionInterface &
  Node & {
    __typename?: "FilterArtworksConnection"
    aggregations?: Maybe<Array<Maybe<ArtworksAggregationResults>>>
    counts?: Maybe<FilterArtworksCounts>
    edges?: Maybe<Array<Maybe<FilterArtworksEdge>>>
    facet?: Maybe<ArtworkFilterFacet>
    /** @deprecated Prefer to use `edges`. [Will be removed in v2] */
    hits?: Maybe<Array<Maybe<Artwork>>>
    id: Scalars["ID"]["output"]
    merchandisableArtists?: Maybe<Array<Maybe<Artist>>>
    pageCursors: PageCursors
    pageInfo: PageInfo
  }
// ...
export type LocationConnection = {
  __typename?: "LocationConnection"
  edges?: Maybe<Array<Maybe<LocationEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type Node = {
  id: Scalars["ID"]["output"]
}
// ...
export type Partner = Node & {
  __typename?: "Partner"
  alertsConnection?: Maybe<PartnerAlertsConnection>
  allArtistsConnection?: Maybe<ArtistPartnerConnection>
  analytics?: Maybe<AnalyticsPartnerStats>
  articlesConnection?: Maybe<ArticleConnection>
  artistsConnection?: Maybe<ArtistPartnerConnection>
  artistsSearchConnection?: Maybe<ArtistConnection>
  artistsWithAlertCountsConnection?: Maybe<ArtistsWithAlertCountsConnection>
  artworkImportsConnection?: Maybe<ArtworkImportConnection>
  artworksConnection?: Maybe<ArtworkConnection>
  artworksSearchConnection?: Maybe<ArtworkConnection>
  cached?: Maybe<Scalars["Int"]["output"]>
  categories?: Maybe<Array<Maybe<PartnerCategory>>>
  cities?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>
  claimed?: Maybe<Scalars["Boolean"]["output"]>
  collectingInstitution?: Maybe<Scalars["String"]["output"]>
  contact?: Maybe<Contact>
  contactsConnection?: Maybe<ContactConnection>
  counts?: Maybe<PartnerCounts>
  defaultProfileID?: Maybe<Scalars["String"]["output"]>
  displayArtistsSection?: Maybe<Scalars["Boolean"]["output"]>
  displayFullPartnerPage?: Maybe<Scalars["Boolean"]["output"]>
  displayWorksSection?: Maybe<Scalars["Boolean"]["output"]>
  distinguishRepresentedArtists?: Maybe<Scalars["Boolean"]["output"]>
  documentsConnection?: Maybe<PartnerDocumentConnection>
  featuredKeywords: Array<Scalars["String"]["output"]>
  featuredShow?: Maybe<Show>
  filterArtworksConnection?: Maybe<FilterArtworksConnection>
  /** @deprecated Prefer displayFullPartnerPage */
  fullProfileEligible?: Maybe<Scalars["Boolean"]["output"]>
  hasFairPartnership?: Maybe<Scalars["Boolean"]["output"]>
  hasVisibleFollowsCount: Scalars["Boolean"]["output"]
  href?: Maybe<Scalars["String"]["output"]>
  id: Scalars["ID"]["output"]
  initials?: Maybe<Scalars["String"]["output"]>
  inquiryRequest?: Maybe<PartnerInquiryRequest>
  internalID: Scalars["ID"]["output"]
  isDefaultProfilePublic?: Maybe<Scalars["Boolean"]["output"]>
  isInquireable: Scalars["Boolean"]["output"]
  isLinkable?: Maybe<Scalars["Boolean"]["output"]>
  isPreQualify?: Maybe<Scalars["Boolean"]["output"]>
  isVerifiedSeller?: Maybe<Scalars["Boolean"]["output"]>
  location?: Maybe<Location>
  /** @deprecated Prefer to use `locationsConnection`. [Will be removed in v2] */
  locations?: Maybe<Array<Maybe<Location>>>
  locationsConnection?: Maybe<LocationConnection>
  merchantAccount?: Maybe<PartnerMerchantAccount>
  meta?: Maybe<PartnerMeta>
  name?: Maybe<Scalars["String"]["output"]>
  partnerAlertHitsConnection?: Maybe<PartnerAlertHitsConnection>
  partnerPageEligible?: Maybe<Scalars["Boolean"]["output"]>
  partnerType?: Maybe<Scalars["String"]["output"]>
  profile?: Maybe<Profile>
  profileArtistsLayout?: Maybe<Scalars["String"]["output"]>
  profileBannerDisplay?: Maybe<Scalars["String"]["output"]>
  showPromoted?: Maybe<Scalars["Boolean"]["output"]>
  showsConnection?: Maybe<ShowConnection>
  showsSearchConnection?: Maybe<ShowConnection>
  slug: Scalars["ID"]["output"]
  type?: Maybe<Scalars["String"]["output"]>
  vatInformation?: Maybe<Scalars["String"]["output"]>
  vatNumber?: Maybe<Scalars["String"]["output"]>
  vatStatus?: Maybe<Scalars["String"]["output"]>
  viewingRoomsConnection?: Maybe<ViewingRoomsConnection>
  website?: Maybe<Scalars["String"]["output"]>
}
// ...
export type PartnerAlertHitsConnection = {
  __typename?: "PartnerAlertHitsConnection"
  edges?: Maybe<Array<Maybe<PartnerAlertHitsEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type PartnerAlertsConnection = {
  __typename?: "PartnerAlertsConnection"
  edges?: Maybe<Array<Maybe<PartnerAlertsEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type PartnerCategory = {
  __typename?: "PartnerCategory"
  cached?: Maybe<Scalars["Int"]["output"]>
  categoryType?: Maybe<PartnerCategoryType>
  id: Scalars["ID"]["output"]
  internal?: Maybe<Scalars["Boolean"]["output"]>
  internalID: Scalars["ID"]["output"]
  name?: Maybe<Scalars["String"]["output"]>
  partners?: Maybe<Array<Maybe<Partner>>>
  slug: Scalars["ID"]["output"]
}
// ...
export type PartnerCounts = {
  __typename?: "PartnerCounts"
  artistDocuments?: Maybe<Scalars["FormattedNumber"]["output"]>
  artists?: Maybe<Scalars["FormattedNumber"]["output"]>
  artworks?: Maybe<Scalars["FormattedNumber"]["output"]>
  currentDisplayableShows?: Maybe<Scalars["FormattedNumber"]["output"]>
  displayableShows?: Maybe<Scalars["FormattedNumber"]["output"]>
  eligibleArtworks?: Maybe<Scalars["FormattedNumber"]["output"]>
  partnerArtists?: Maybe<Scalars["FormattedNumber"]["output"]>
  partnerShowDocuments?: Maybe<Scalars["FormattedNumber"]["output"]>
  publishedForSaleArtworks?: Maybe<Scalars["FormattedNumber"]["output"]>
  publishedNotForSaleArtworks?: Maybe<Scalars["FormattedNumber"]["output"]>
  shows?: Maybe<Scalars["FormattedNumber"]["output"]>
}
// ...
export type PartnerDocumentConnection = {
  __typename?: "PartnerDocumentConnection"
  edges?: Maybe<Array<Maybe<PartnerDocumentEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type PartnerInquiryRequest = {
  __typename?: "PartnerInquiryRequest"
  collectorProfile?: Maybe<InquirerCollectorProfile>
  formattedFirstMessage?: Maybe<Scalars["String"]["output"]>
  id: Scalars["ID"]["output"]
  internalID: Scalars["ID"]["output"]
  questions?: Maybe<Array<Maybe<InquiryQuestion>>>
  shippingLocation?: Maybe<Location>
}

export type PartnerMerchantAccount = {
  __typename?: "PartnerMerchantAccount"
  externalId: Scalars["String"]["output"]
}

export type PartnerMeta = {
  __typename?: "PartnerMeta"
  description?: Maybe<Scalars["String"]["output"]>
  image?: Maybe<Scalars["String"]["output"]>
  title?: Maybe<Scalars["String"]["output"]>
}
// ...
export type ShowConnection = {
  __typename?: "ShowConnection"
  edges?: Maybe<Array<Maybe<ShowEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type ViewingRoomsConnection = {
  __typename?: "ViewingRoomsConnection"
  edges?: Maybe<Array<Maybe<ViewingRoomsEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type InquiriesStatsToolQuery = {
  __typename?: "Query"
  partner?: {
    __typename?: "Partner"
    name?: string | null
    analytics?: {
      __typename?: "AnalyticsPartnerStats"
      inquiry?: {
        __typename?: "AnalyticsPartnerInquiryStats"
        inquiryCount: number
        inquiryResponseTime?: number | null
        timeSeries?: Array<{
          __typename?: "AnalyticsPartnerInquiryCountTimeSeriesStats"
          count?: number | null
          startTime?: any | null
          endTime?: any | null
        }> | null
      } | null
    } | null
  } | null
}